requirements.txt
openai-agents-python
google-generativeai
azure-identity
azure-mgmt-resource
azure-mgmt-keyvault
typer
streamlit
python-dotenv
